+++ src/Parser/JavaScript/JavaScriptTokenKind.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Symfony\Lsp\Parser\JavaScript;
+
+enum JavaScriptTokenKind
+{
+    case Identifier;
+    case Number;
+    case String;
+    case Template;
+    case RegularExpression;
+    case Punctuator;
+    case Comment;
+}